requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.0
twilio==8.10.0
python-dotenv==1.0.0
pandas==2.1.3
openpyxl==3.1.2
colorama==0.4.6
tabulate==0.9.0
schedule==1.2.1
jinja2==3.1.2
lxml==4.9.3
